fixed purgeDiskCache

+++ library/src/uk/co/senab/bitmapcache/BitmapLruCache.java
@@ -16,8 +16,6 @@
 
 package uk.co.senab.bitmapcache;
 
-import com.jakewharton.disklrucache.DiskLruCache;
-
 import android.content.Context;
 import android.content.res.Resources;
 import android.graphics.Bitmap;
@@ -27,6 +25,8 @@ import android.os.Build;
 import android.os.Looper;
 import android.os.Process;
 import android.util.Log;
+
+import com.jakewharton.disklrucache.DiskLruCache;
 
 import java.io.ByteArrayInputStream;
 import java.io.File;
@@ -652,10 +652,13 @@ public class BitmapLruCache {
     }
 
     public void purgeDiskCache() {
-        try {
-            mDiskCache.delete();
-        } catch (IOException ex) {
-            ex.printStackTrace();
+        if (null != mDiskCache) {
+            checkNotOnMainThread();
+            try {
+                mDiskCache.delete();
+            } catch (IOException ex) {
+                ex.printStackTrace();
+            }
         }
     }
